Set HTML and CSS files
In the HTML file the header was set with a DIV as well as a demo card
for the books to be displayed.

In the CSS file the header font was set by downloading it from google.
The color of the header was also set with the help of a variable.