Style header with SVG wave and background image
The image was placed in CSS with the background declaration. Likewise,
the opacity was given by adding an extra DIV in the header section in
HTML, which was given an absolute position as well as the SVG wave.

Finally, the title was given z-index 1 to place it above the opacity.