Style display for books and apply grid to arrange
The display for the books were made by rotating the DIV containing the
title 20deg in the x-axis. Likewise, it was added a pseudo-element to
give the appearance of the thick of pages at the bottom, as the book
gives the perspective of being seen from the bottom.

On the other hand, the container for the books was defined as grid,
so with the help of auto-fit and the minmax() function any amount of
books can be placed.