Add background with opacity for the pop up form
The element for the pop up form was set in the HTML file afther the
container for the books. It was set an absolute position to remove it
from the flow, so it would not affect other elements.

Within this new element the DIV for the form was defined as well as
three inputs to specify the title, author, and number of pages of the
book on the side of the user.