Give function to 'Add new book' and 'x' buttons
The 'Add new book' button opens the form so the user can enter the data
of the book. On the other hand, if by accident the user opens the form,
it is possible to close it by means of the 'x' button in the top right
part of the form.