Extract information from form
The variables containing the corresponding inputs were declared as
global to manipulate them outside the scope of the function that
extracts the information.

The element of the form was given an event listener "submit" that calls
the function collectDataForm that updates the values of the inputs and
also prevents the submit button from refreshing the page.