Create function createBook()
The function creates 6 DIV elements and two buttons, each one receiving
a class to have the style created with the demo card. Once this is done,
the function appendChild() is used to start nesting every element
according to the desired structure in HTML, resulting in the creation of
the book on the webpage.