Add toggle function to button Read
This was made possible by returning the variable containing the element
of the button, which was added an event listener to call the function
that compare the actual class of the element to decide what the
resulting state should be.

In the Css file a blue color was set to the button when the book has not
been read.