Add  funcionality to button 'remove'
This was made by setting an ID to the container of the corresponding
book as well as an extra class for the 'remove' button. Thus, when
the event listener was called it would select the element that matches
with the number of class of the button. Therefore, the container was
selected and all of its child elements could be erased as well as the
container itself.